chore(rust): move `SymbolOrMemberExprRef` and `MemberExprRef` to seperate files (#1816)
<!-- Thank you for contributing! -->

### Description

<!-- Please insert your description here and provide especially info about the "what" this PR is solving -->
