+++ crates/rolldown_common/src/types/mod.rs
@@ -13,6 +13,7 @@ pub mod external_module_idx;
 pub mod import_record;
 pub mod importer_record;
 pub mod js_regex;
+pub mod member_expr_ref;
 pub mod module_def_format;
 pub mod module_id;
 pub mod module_idx;
@@ -30,5 +31,6 @@ pub mod rollup_rendered_chunk;
 pub mod side_effects;
 pub mod stmt_info;
 pub mod str_or_bytes;
+pub mod symbol_or_member_expr_ref;
 pub mod symbol_ref;
 pub mod wrap_kind;
